Refine IntervallicFragmentable organization, doc comments
